moved dao spesific methods to dao folder
